HillelBookShop/requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
dj-database-url==3.0.1
Django==4.2.28
django-crispy-forms==2.5
django-debug-toolbar==6.1.0
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-mptt==0.18.0
django-widget-tweaks==1.5.1
gunicorn==23.0.0
packaging==26.0
pillow==11.3.0
psycopg-binary==3.2.13
psycopg-pool==3.2.8
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
